% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Agric_NSW_GMNarrendera}
\alias{Agric_NSW_GMNarrendera}
\title{Farm Adaptation at Narrendera, New South Wales}
\format{
csv file with 50 rows and 5 columns
}
\source{
https://nccarf.edu.au/will-primary-producers-continue-adjust-practices-and-technologies-change-production/
}
\description{
Will primary producers continue to adjust practices and technologies, change production systems or transform their industry – An application of real options.
}
\details{
\itemize{
\item Year: time variable in annual increments
\item CWWCWW GM: Gross margins per hectare for CWWCWW rotation
\item CWWSSS GM: Gross margins per hectare for CWWSSS rotation
\item SSSSC GM: Gross margins per hectare for SSSSC rotation
\item S GM: Gross margins per hectare for continuous S
}

APSIM generated gross margins for wheat (W), canola (C) and sheep (S).
}
\author{
Greg Hertzler, Todd Sanderson, Tim Capon, Peter Hayman, Ross Kingwell, Anthea McClintock, Jason Crean and Alan Randall, 2013
}
\keyword{datasets}
