% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Agric_NSW_SoilHealthHarden}
\alias{Agric_NSW_SoilHealthHarden}
\title{Soil Health with Stubble Management}
\format{
csv file with 30 rows and 5 columns
}
\source{
https://doi.org/10.25919/2jqe-mz45
}
\description{
CSIRO Harden Long-Term Tillage Experiment. v3. CSIRO. Data Collection.
}
\details{
\itemize{
\item Year: time variable in sporadic increments
\item N Burn: kilograms per hectare of mineral soil nitrogen in the soil profile from 0 to 160 millimetres with stubble burning
\item N Bash: kilograms per hectare of mineral soil nitrogen in the soil profile from 0 to 160 millimetres with stubble bashing
\item W Burn: millimetres of water in the soil profile from 0 to 160 millimetres with stubble burning
\item W Bash: millimetres of water in the soil profile from 0 to 160 millimetres with stubble bashing
}

Soil nitrogen and water in the top 160 millimetres.
}
\author{
Kirkegaard, John; & Lilley, Julianne, 2023
}
\keyword{datasets}
