% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Agric_SA_GMOrroroo}
\alias{Agric_SA_GMOrroroo}
\title{Farm Adaptation at Orroroo, South Australia}
\format{
csv file with 108 rows and 7 columns
}
\source{
https://nccarf.edu.au/will-primary-producers-continue-adjust-practices-and-technologies-change-production/
}
\description{
Will primary producers continue to adjust practices and technologies, change production systems or transform their industry – An application of real options.
}
\details{
\itemize{
\item Year: time variable in annual increments
\item Annual rain: Annual rainfall in millimetres
\item Apr-Oct rain: Growing season rainfall in millimetres
\item Wheat Yld: Wheat yield in tonnes per hectare
\item Sheep DSE: Dry-sheep equivalents per hectare
\item Wheat GM: Wheat gross margins per hectare
\item Sheep GM: Sheep gross margins per hectare
}

APSIM generated Wheat yields and manually simulated dry-sheep equivalents.
}
\author{
Greg Hertzler, Todd Sanderson, Tim Capon, Peter Hayman, Ross Kingwell, Anthea McClintock, Jason Crean and Alan Randall, 2013
}
\keyword{datasets}
