% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Agric_Tas_TreeShelterBelts}
\alias{Agric_Tas_TreeShelterBelts}
\title{Tree shelter belts in Tasmania}
\format{
csv file with 52 rows and 3 columns
}
\source{
corresponding author, Tim Capon \email{tim.capon@csiro.au}
}
\description{
CSIRO Perennial Prosperity Project.
}
\details{
\itemize{
\item Year: time variable in annual increments
\item With trees: sheep gross margins per paddock with tree shelter belt
\item Without trees: sheep gross margins per paddock without tree shelter belt
}

Simulated effects of shelter belts with prices and costs for sheep production.
}
\author{
Tim Capon, Daniel Mendham, 2024
}
\keyword{datasets}
