% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Climate_CarbonCredits_EECXM}
\alias{Climate_CarbonCredits_EECXM}
\title{Price of Carbon}
\format{
csv file with 1519 rows and 6 columns
}
\source{
https://seekingalpha.com/symbol/EECXM
}
\description{
European Union Emission Trading System
}
\details{
\itemize{
\item Year: time variable as decimal year
\item Open: Price at market open in EUR/tonne
\item High: Daily high price in EUR/tonne
\item Low: Daily low price in EUR/tonne
\item Close: Price at market close in EUR/tonne
\item Day: time variable in days since 1 January 2021
}

Daily price of Carbon Permits from 1 January 2021 to 31 December 2025
}
\keyword{datasets}
