% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Climate_TempsMax_Darwin}
\alias{Climate_TempsMax_Darwin}
\title{Maximum daily temperatures at Darwin}
\format{
csv file with 9127 rows and 2 columns
}
\source{
http://www.bom.gov.au/climate/data/?ref=ftr, Station number 014040
}
\description{
Climate Data Online
}
\details{
\itemize{
\item Year: time variable in daily increments
\item Max: maximum temperature in degrees centigrade
}

Raw (Unhomogenized) data.
}
\author{
Bureau of Meteorology
}
\keyword{datasets}
