% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Default}
\alias{Default}
\title{Default data}
\format{
csv file with 6 rows and 2 columns
}
\description{
Simulated data to estimate default parameters, rho=0.5, mu=15 and sigma=15,
where rho is the rate of convergence, mu is the location and sigma is the
scale.
}
\details{
\itemize{
\item tau: time variable
\item z: state variable
}
}
\keyword{datasets}
