% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Ecosys_IrrigationWaterSupply}
\alias{Ecosys_IrrigationWaterSupply}
\title{Water Supply for Irrigated Agriculture in Eastern Australia}
\format{
csv file with 13 rows and 2 columns
}
\source{
https://doi.org/10.25919/7jj7-8826
}
\description{
National data for Australia's ecosystem services: irrigated agricultural water supply (250 m resolution): 2000-01 to 2022-23. v7. CSIRO. Data Collection.
}
\details{
\itemize{
\item Fin Year: time variable for financial year 1 July to 30 June
\item WaterSupply: irrigation water supplied in Australia in megalitres
}

Regions are aggregated for Eastern Australia.
}
\author{
Liu, Ning; Smith, Greg; Evans, David; Tetreault Campbell, Sally; Pascoe, Sean; & Schmidt, Becky, 2024
}
\keyword{datasets}
