% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Ecosys_Kangaroos}
\alias{Ecosys_Kangaroos}
\title{Kangaroo Population and Harvest in South Australia}
\format{
csv file with 46 rows and 7 columns
}
\source{
https://www.environment.sa.gov.au/topics/animals-and-plants/sustainable-use-of-animals-and-plants/kangaroo-conservation-and-management/quotas-harvest-data
}
\description{
Population surveys and meat processor records of harvest.
}
\details{
\itemize{
\item Year: time variable
\item Red Pop: Estimated population of Red Kangaroos in 1000 head
\item Red Harvest: Number of Red Kangaroos commercially harvested in 1000 head
\item Grey Pop: Estimated population of Western Grey Kangaroos in 1000 head
\item Grey Harvest: Number of Western Grey Kangaroos commercially harvested in 1000 head
\item Euro Pop: Estimated population of European Kangaroos in 1000 head
\item Euro Harvest: Number of European Kangaroos commercially harvested in 1000 head
}

Regions are aggregated for all of South Australia.
}
\author{
Government of South Australia, Department of Environment
}
\keyword{datasets}
