% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Finance_Commodities}
\alias{Finance_Commodities}
\title{Metals and Energy Commodities}
\format{
csv file with 197 rows and 8 columns
}
\source{
https://files.marketindex.com.au/files/workbooks/commodities-workbook.xlsx
}
\description{
London Bullion Market Association, World Gold Council, GoldHub, Perth Mint and International Monetary Fund prices.
}
\details{
\itemize{
\item Day: time variable in days since 1 January 2009
\item Gold: London Bullion Market fix at the Perth (Aus) Mint in USD/troy ounce
\item Silver: London Bullion Market fix in USD/troy ounce
\item Copper: London Metals Exchange spot price, Copper, Grade A, USD/tonne
\item Iron Ore: Cleared for Export Tianjin port, Iron Ore Fines 62\% FE spot price in USD/tonne
\item WTI: West Texas Intermediate in USD/bbl
\item Brent: Brent Crude in USD/bbl
\item Year: time variable as decimal year
}

Closing prices on the first day of each month from 1 January 2009 to 1 May 2025.
}
\keyword{datasets}
