% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{Finance_KansasCity_WheatFutures}
\alias{Finance_KansasCity_WheatFutures}
\title{Futures Prices}
\format{
csv file with 250 rows and 8 columns
}
\source{
https://finance.yahoo.com/quote (search for KE=F)
}
\description{
Kansas City Hard Red Wheat Futures
}
\details{
\itemize{
\item Day: time variable in days since 1 July 2024
\item Open: Price at market open in USD/ton
\item High: Daily high price in USD/ton
\item Low: Daily low price in USD/ton
\item Close: Price at market close in USD/ton
\item Adj Close: Same as Close for futures
\item Volume: Daily number of trades
\item Year: time variable as decimal year
}

Daily futures for September 2025 maturity from 1 July 2024 to 30 June 2025
}
\keyword{datasets}
