% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{GregsOUPR6-package}
\alias{GregsOUPR6}
\alias{GregsOUPR6-package}
\title{GregsOUPR6: Real Options for Adoption and Resilience: Ornstein-Uhlenbeck Process: R6 Objects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A complete set of functions for maximum likelihood estimation and for the calculation of probabilities, option prices, decision thresholds, visiting times, first passage times and more--everything for a real options analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://greghertzler.github.io/GregsOUPR6/}
  \item \url{https://greghertzler.github.io/GregsOUPR6/OUP_Help.html}
  \item \url{https://greghertzler.github.io/GregsOUPShiny/OUP_Shiny.html}
  \item \url{https://real-options-adoption-resilience.shinyapps.io/GregsOUPShiny/}
}

}
\author{
\strong{Maintainer}: Greg Hertzler \email{ghertzlerau@gmail.com} (\href{https://orcid.org/0000-0003-3123-7898}{ORCID})

Authors:
\itemize{
  \item Greg Hertzler \email{ghertzlerau@gmail.com} (\href{https://orcid.org/0000-0003-3123-7898}{ORCID})
}

Other contributors:
\itemize{
  \item Tim Capon [contributor]
  \item CSIRO IMT Scientific Computing [contributor, funder]
  \item CSIRO Environment [funder]
}

}
\keyword{internal}
