% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\name{MaximumLikelihood}
\alias{MaximumLikelihood}
\title{R6 class implementing Maximum Likelihood estimation of the Ornstein-Uhlenbeck Process.}
\description{
Maximum Likelihood estimation using a modified Nelder-Mead algorithm with testing
for simple hypotheses.
}
\section{Formulas and Methods:}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{rho, mu and sigma random
  LogLikelihood
  Estimates
  GoodnessOfFit
  LikelihoodRatioTest
}\if{html}{\out{</div>}}
}

\section{Plots:}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  PlotTimeSeries
  PlotEstimates
}\if{html}{\out{</div>}}
}

\section{Arguments of functions:}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  All arguments are optional in all functions.
  rho:    rate parameter 0<=rho<inf
  mu:     location parameter -inf<mu<inf
  sigma:  scale parameter -inf<sigma<inf
  tau:    vector of observed times -inf<tau<inf
  z:      vector of observed states -inf<z<inf
  df:     data frame containing columns for tau and z
  taucol: index of a column containing tau
  zcol:   index of a column containing z
  rhor:   constant to fix the rate parameter 0<=rhor<inf
  mur:    constant to fix the location parameter -inf<mur<inf
  sigmar: constant to fix scale parameter -inf<sigmar<inf
  rhos:   starting value for the rate parameter 0<=rhos<inf
  mus:    starting value for the location parameter -inf<mus<inf
  sigmas: starting value the scale parameter -inf<sigmas<inf
  lnLu:   unrestricted log likelihood -inf<lnLu<=0
  alphau: identifies distribution of lnLu 0<=alphau<=1
  lnLr:   restricted log likelihood -inf<lnLr<=lnLu
  alphar: identifies distribution of lnLr 0<=alphar<=1
  m1:     number of observations 0<m1=m-1
}\if{html}{\out{</div>}}
}

\section{Usage:}{
The MaximumLikelihood object must first be instantiated before its methods
are called.  There are two ways.  The first way instantiates the OUProcess
object and calls a function to get a pointer:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  OUP <- OUProcess$new()
  A <- OUP$get_Analytical()
  FD <- OUP$get_FiniteDifference()
  ML <- OUP$get_MaximumLikelihood()
  MC <- OUP$get_MonteCarlo()
}\if{html}{\out{</div>}}

The MaximumLikelihood object will coordinate arguments to functions with the
Analytical, FiniteDifference and MonteCarlo objects.  The second way
instantiates the MaximumLikelihood object by itself with no coordination:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  ML <- MaximumLikelihood$new()
}\if{html}{\out{</div>}}

Once the object is instantiated, its methods can be called, to estimate the
parameters of the Ornstein-Uhlenbeck Process, for example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  ML$Estimate()
}\if{html}{\out{</div>}}

The plot methods can be used to customize the plots, with a title, for example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  ML$PlotEstimates(title="My Estimates")
}\if{html}{\out{</div>}}

Other functions and methods are called in the same way.  To see all the
possibilities and, in particular, how to read in data, check out the demos below.
}

\section{Demos:}{
Demonstration scripts are in files in the 'demo' directory. The most
convenient way to list and run demos are the commands:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  OUPDemoList()
  OUPDemoRun(<number of demo in list>)
}\if{html}{\out{</div>}}

Entering the demos by number in the list saves typing.
}

\section{Discussion:}{
The Nelder-Mead algorithm is used to maximize a Log Likelihood derived from
the Transition Density of the Ornstein-Uhlenbeck Process. The original
algorithm is designed to minimize and has been called the amoeba algorithm.
If spread across a bumpy surface, the amoeba pulls itself over bumps and out
of hollows and flows down to the lowest level.  There it contracts around
its center.  If spread across a flat surface, the amoeba shrinks without
flowing.  Once the amoeba contracts or shrinks, it is spread across the
surface opposite from where it came.  Again it flows, contracts and shrinks.
Spreading, flowing, contracting and shrinking continue until the amoeba
contracts around the same point or shrinks at the same level at least twice.

In this implementation, the Nelder-Mead algorithm has been modified in five
ways.
\enumerate{
\item It is modified to maximize;
\item It can set parameters rho, mu and/or sigma to constants for simple
hypothesis tests;
\item It has a tie-breaking condition to prevent cycling or freezing;
\item It can accelerate over long distances for searching flat likelihoods;
\item It checks for log likelihoods greater than zero to prevent errors in
case the data is faked or artificially smoothed.
}

Once the unrestricted estimates are found, the arguments for restricting the
parameters can be set for simple hypothesis tests. By default, starting
values are calculated from the data and can be updated automatically. The
arguments for starting values should seldom be needed.

Data are entered as a data frame with columns for tau, a vector of times, and
z, a vector of states. Arguments taucol and zcol are columns to extract from
a data frame containing many columns.

To read a csv file into a data frame, create a MaximumLikelihood object,
estimate parameters and test the goodness-of-fit:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  df <- read.csv("data/MyData.csv")
  ML <- MaximumLikelihood$new()
  ML$Estimates(df)
  ML$GoodnessOfFit()
}\if{html}{\out{</div>}}

In this example, the 'data' directory is in the working directory. The taucol
and zcol arguments are optional and will default to 1 and 2. Continue on by
restricting a parameter and performing a likelihood ratio test:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  ML$Estimates(rhor=0.5)
  ML$LikelihoodRatioTest()
}\if{html}{\out{</div>}}

You can also list and read from available data sets using:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ OUPDataList()
 df <- OUPDataRead(27)
}\if{html}{\out{</div>}}

Number 27 in the list is the data set "OUP_Convergence".  The available data
sets are documented. Type one of the following categories:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  ?Agric_
  ?Climate_
  ?Ecosys_
  ?Finance_
  ?OUP_
}\if{html}{\out{</div>}}

and then select from the drop-down list.  Or use the function:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  OUPDataHelp(27)
}\if{html}{\out{</div>}}
}

\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-MaximumLikelihood-initialize}{\code{MaximumLikelihood$new()}}
    \item \href{#method-MaximumLikelihood-set_oup_params}{\code{MaximumLikelihood$set_oup_params()}}
    \item \href{#method-MaximumLikelihood-set_oup_params_restr}{\code{MaximumLikelihood$set_oup_params_restr()}}
    \item \href{#method-MaximumLikelihood-set_oup_params_start}{\code{MaximumLikelihood$set_oup_params_start()}}
    \item \href{#method-MaximumLikelihood-set_timeseries}{\code{MaximumLikelihood$set_timeseries()}}
    \item \href{#method-MaximumLikelihood-set_timeseries_info}{\code{MaximumLikelihood$set_timeseries_info()}}
    \item \href{#method-MaximumLikelihood-set_plot_info}{\code{MaximumLikelihood$set_plot_info()}}
    \item \href{#method-MaximumLikelihood-set_flags}{\code{MaximumLikelihood$set_flags()}}
    \item \href{#method-MaximumLikelihood-get_all}{\code{MaximumLikelihood$get_all()}}
    \item \href{#method-MaximumLikelihood-get_oup_params}{\code{MaximumLikelihood$get_oup_params()}}
    \item \href{#method-MaximumLikelihood-get_oup_params_restr}{\code{MaximumLikelihood$get_oup_params_restr()}}
    \item \href{#method-MaximumLikelihood-get_oup_params_start}{\code{MaximumLikelihood$get_oup_params_start()}}
    \item \href{#method-MaximumLikelihood-get_timeseries}{\code{MaximumLikelihood$get_timeseries()}}
    \item \href{#method-MaximumLikelihood-get_timeseries_info}{\code{MaximumLikelihood$get_timeseries_info()}}
    \item \href{#method-MaximumLikelihood-get_plot_info}{\code{MaximumLikelihood$get_plot_info()}}
    \item \href{#method-MaximumLikelihood-get_plot_colors}{\code{MaximumLikelihood$get_plot_colors()}}
    \item \href{#method-MaximumLikelihood-get_flags}{\code{MaximumLikelihood$get_flags()}}
    \item \href{#method-MaximumLikelihood-LogLikelihood}{\code{MaximumLikelihood$LogLikelihood()}}
    \item \href{#method-MaximumLikelihood-Estimates}{\code{MaximumLikelihood$Estimates()}}
    \item \href{#method-MaximumLikelihood-GoodnessOfFit}{\code{MaximumLikelihood$GoodnessOfFit()}}
    \item \href{#method-MaximumLikelihood-LikelihoodRatioTest}{\code{MaximumLikelihood$LikelihoodRatioTest()}}
    \item \href{#method-MaximumLikelihood-PlotTimeSeries}{\code{MaximumLikelihood$PlotTimeSeries()}}
    \item \href{#method-MaximumLikelihood-PlotEstimates}{\code{MaximumLikelihood$PlotEstimates()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-initialize}{}}}
\subsection{\code{MaximumLikelihood$new()}}{
  Create a MaximumLikelihood object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$new(OUP = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{OUP}}{pointer set by the OUProcess object}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new MaximumLikelihood object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-set_oup_params"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-set_oup_params}{}}}
\subsection{\code{MaximumLikelihood$set_oup_params()}}{
  Set OUP parameters
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$set_oup_params(
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  who = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rho}}{rate parameter 0<=rho<inf}
      \item{\code{mu}}{location parameter -inf<mu<inf}
      \item{\code{sigma}}{scale parameter -inf<sigma<inf}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(rho,mu,sigma)
Set OUP parameter restrictions
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-set_oup_params_restr"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-set_oup_params_restr}{}}}
\subsection{\code{MaximumLikelihood$set_oup_params_restr()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$set_oup_params_restr(rhor = NULL, mur = NULL, sigmar = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rhor}}{rate parameter 0<=rhor<inf}
      \item{\code{mur}}{location parameter -inf<mur<inf}
      \item{\code{sigmar}}{scale parameter -inf<sigmar<inf}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(rhor,mur,sigmar)
Set OUP parameter starting values
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-set_oup_params_start"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-set_oup_params_start}{}}}
\subsection{\code{MaximumLikelihood$set_oup_params_start()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$set_oup_params_start(rhos = NULL, mus = NULL, sigmas = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rhos}}{rate parameter 0<=rhos<inf}
      \item{\code{mus}}{location parameter -inf<mus<inf}
      \item{\code{sigmas}}{scale parameter -inf<sigmas<inf}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(rhos,mus,sigmas)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-set_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-set_timeseries}{}}}
\subsection{\code{MaximumLikelihood$set_timeseries()}}{
  Set time series data for time tau and state z
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$set_timeseries(df = NULL, taucol = NULL, zcol = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{df}}{data frame containing columns for tau and z}
      \item{\code{taucol}}{index of a column containing tau}
      \item{\code{zcol}}{index a column containing z}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    dataframe(tau,z)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-set_timeseries_info"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-set_timeseries_info}{}}}
\subsection{\code{MaximumLikelihood$set_timeseries_info()}}{
  Set information for plotting times series and estimates
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$set_timeseries_info(
  tbeg = NULL,
  tend = NULL,
  dataname = NULL,
  timename = NULL,
  statename = NULL,
  estimation = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{tbeg}}{begin value for time axis}
      \item{\code{tend}}{end value for time axis}
      \item{\code{dataname}}{name for the data}
      \item{\code{timename}}{name for time}
      \item{\code{statename}}{name for state}
      \item{\code{estimation}}{description of estimation}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(tbeg,Ixbeg,tend,Ixend,datename,timename,statename,estimation)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-set_plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-set_plot_info}{}}}
\subsection{\code{MaximumLikelihood$set_plot_info()}}{
  Set information for plotting
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$set_plot_info(
  fontfamily = NULL,
  fontsize = NULL,
  fileformat = NULL,
  filewidth = NULL,
  fileheight = NULL,
  theme = NULL,
  opaque = NULL,
  labels = NULL,
  who = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fontfamily}}{font family for plot labels}
      \item{\code{fontsize}}{font size for plot labels}
      \item{\code{fileformat}}{'png' or 'svg'}
      \item{\code{filewidth}}{pixel width of 2D plot, pixel width and height of 3D plot}
      \item{\code{fileheight}}{pixel height of 2D plot}
      \item{\code{theme}}{'light' or 'dark'}
      \item{\code{opaque}}{transparent to opaque background 0.0<=opaque<=1.0}
      \item{\code{labels}}{title and parameters TRUE or FALSE}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(font,file,theme,labels)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-set_flags"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-set_flags}{}}}
\subsection{\code{MaximumLikelihood$set_flags()}}{
  Set flags for plotting and copying
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$set_flags(plotit = NULL, copyit = NULL, who = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plotit}}{automatic plot after calculation TRUE or FALSE}
      \item{\code{copyit}}{copy to clipboard TRUE or FALSE}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(plotit,copyit)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_all"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_all}{}}}
\subsection{\code{MaximumLikelihood$get_all()}}{
  Get all arguments, time series and information
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_all()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(oup_params,oup_params_restr,oup_params_start,timeseries,timeseries_info,plot_info)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_oup_params"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_oup_params}{}}}
\subsection{\code{MaximumLikelihood$get_oup_params()}}{
  Get OUP parameters
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_oup_params()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(rho,mu,sigma)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_oup_params_restr"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_oup_params_restr}{}}}
\subsection{\code{MaximumLikelihood$get_oup_params_restr()}}{
  Get OUP parameter restrictions
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_oup_params_restr()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(rhor,mur,sigmar)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_oup_params_start"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_oup_params_start}{}}}
\subsection{\code{MaximumLikelihood$get_oup_params_start()}}{
  Get OUP parameter starting values
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_oup_params_start()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(rhos,mus,sigmas)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_timeseries}{}}}
\subsection{\code{MaximumLikelihood$get_timeseries()}}{
  Get time series data for time tau and state z
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_timeseries()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    dataframe(tau,z)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_timeseries_info"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_timeseries_info}{}}}
\subsection{\code{MaximumLikelihood$get_timeseries_info()}}{
  Get information for times series
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_timeseries_info()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(tbeg,tend,datename,timename,statename,estimation)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_plot_info}{}}}
\subsection{\code{MaximumLikelihood$get_plot_info()}}{
  Get information for plotting
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_plot_info()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(font,file,theme,labels)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_plot_colors"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_plot_colors}{}}}
\subsection{\code{MaximumLikelihood$get_plot_colors()}}{
  Get colors for plotting
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_plot_colors()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(red,ylw,grn,cyn,blu,mgn,gry,background,font)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-get_flags"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-get_flags}{}}}
\subsection{\code{MaximumLikelihood$get_flags()}}{
  Get flags for plotting and copying
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$get_flags()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(plotit,copyit)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-LogLikelihood"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-LogLikelihood}{}}}
\subsection{\code{MaximumLikelihood$LogLikelihood()}}{
  Calculate the log likelihood of estimates
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$LogLikelihood(
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  df = NULL,
  taucol = NULL,
  zcol = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rho}}{rate parameter 0<=rho<inf}
      \item{\code{mu}}{location parameter -inf<mu<inf}
      \item{\code{sigma}}{scale parameter -inf<sigma<inf}
      \item{\code{df}}{data frame containing columns for tau and z}
      \item{\code{taucol}}{index of a column containing tau}
      \item{\code{zcol}}{index of a column containing z}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(rho,mu,sigma,lnL,k,alpha,m1)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-Estimates"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-Estimates}{}}}
\subsection{\code{MaximumLikelihood$Estimates()}}{
  Calculate unrestricted or restricted maximum likelihood estimates
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$Estimates(
  df = NULL,
  taucol = NULL,
  zcol = NULL,
  rhor = NULL,
  mur = NULL,
  sigmar = NULL,
  rhos = NULL,
  mus = NULL,
  sigmas = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{df}}{data frame containing columns for tau and z}
      \item{\code{taucol}}{index of a column containing tau}
      \item{\code{zcol}}{index of a column containing z}
      \item{\code{rhor}}{constant to fix the rate parameter 0<=rhor<inf}
      \item{\code{mur}}{constant to fix the location parameter -inf<mur<inf}
      \item{\code{sigmar}}{constant to fix scale parameter -inf<sigmar<inf}
      \item{\code{rhos}}{starting value for the rate parameter 0<=rhos<inf}
      \item{\code{mus}}{starting value for the location parameter -inf<mus<inf}
      \item{\code{sigmas}}{starting value the scale parameter -inf<sigmas<inf}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(rhohat,muhat,sigmahat,lnLu,ku,alphau,m1) or list(rhor,mur,sigmar,lnLr,kr,alphar,m1)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-GoodnessOfFit"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-GoodnessOfFit}{}}}
\subsection{\code{MaximumLikelihood$GoodnessOfFit()}}{
  Compare maximum likeliood estimates with invariant and scaled brownian motion estimates
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$GoodnessOfFit(
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  df = NULL,
  taucol = NULL,
  zcol = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rho}}{rate parameter 0<=rho<inf}
      \item{\code{mu}}{location parameter -inf<mu<inf}
      \item{\code{sigma}}{scale parameter -inf<sigma<inf}
      \item{\code{df}}{data frame containing columns for tau and z}
      \item{\code{taucol}}{index of a column containing tau}
      \item{\code{zcol}}{index of a column containing z}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(theta,theta_i,theta_s,Inv,SBM) with Inv=list(R2,Pval) and SBM=list(R2,Pval)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-LikelihoodRatioTest"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-LikelihoodRatioTest}{}}}
\subsection{\code{MaximumLikelihood$LikelihoodRatioTest()}}{
  Test for significant difference between restricted and unrestricted likelihoods
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$LikelihoodRatioTest(
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  df = NULL,
  taucol = NULL,
  zcol = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rho}}{rate parameter 0<=rho<inf}
      \item{\code{mu}}{location parameter -inf<mu<inf}
      \item{\code{sigma}}{scale parameter -inf<sigma<inf}
      \item{\code{df}}{data frame containing columns for tau and z}
      \item{\code{taucol}}{index of a column containing tau}
      \item{\code{zcol}}{index of a column containing z}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    list(theta_u,theta,R2,Pval)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-PlotTimeSeries"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-PlotTimeSeries}{}}}
\subsection{\code{MaximumLikelihood$PlotTimeSeries()}}{
  Plot time series
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$PlotTimeSeries(
  df = NULL,
  taucol = NULL,
  zcol = NULL,
  tbeg = NULL,
  tend = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{df}}{data frame containing columns for tau and z}
      \item{\code{taucol}}{index of a column containing tau}
      \item{\code{zcol}}{index a column containing z}
      \item{\code{tbeg}}{begin value for time axis}
      \item{\code{tend}}{end value for time axis}
      \item{\code{title}}{text for plot title}
      \item{\code{xaxis}}{text for x-axis label}
      \item{\code{yaxis}}{text for y-axis label}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaximumLikelihood-PlotEstimates"></a>}}
\if{latex}{\out{\hypertarget{method-MaximumLikelihood-PlotEstimates}{}}}
\subsection{\code{MaximumLikelihood$PlotEstimates()}}{
  Plot time series with estimated means and variances
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaximumLikelihood$PlotEstimates(
  tbeg = NULL,
  tend = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{tbeg}}{begin value for time axis}
      \item{\code{tend}}{end value for time axis}
      \item{\code{title}}{text for plot title}
      \item{\code{xaxis}}{text for x-axis label}
      \item{\code{yaxis}}{text for y-axis label}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    plot
  }
}

}
