% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_OUP_SimulatedData}
\alias{ML_OUP_SimulatedData}
\title{Simulated sample paths for the Ornstein-Uhlenbeck Process}
\format{
csv file with 177 rows and 16 columns
}
\description{
Monte-Carlo simulation of the stochastic integral equation, with different
rates of convergence.
}
\details{
\itemize{
\item year: time variable in annual increments for all sample paths
\item z1-z5: sample paths in sets of three, each set with the same pseudo-random shocks but different rates of convergence
}
}
\keyword{datasets}
