% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_OUP_UnequalIntervals}
\alias{ML_OUP_UnequalIntervals}
\title{Unequal observation intervals and the Ornstein-Uhlenbeck Process}
\format{
csv file with 151 rows and 4 columns
}
\description{
Monte-Carlo simulation of the stochastic integral equation, with observations
having unequal time intervals
}
\details{
\itemize{
\item year 1: time variable in annual increments
\item z rho 0.5: sample paths with annual time increments or longer and a rate of convergence of 0.5
\item year 0.05: time variable in increments of 1/20 year
\item z rho 10: sample paths with time increments of 1/20 year or longer and a rate of convergence of 10
}
}
\keyword{datasets}
