% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_SeaLevel_PortKembla}
\alias{ML_SeaLevel_PortKembla}
\title{Sea Level at Port Kembla}
\format{
csv file with 7986 rows and 5 columns
}
\source{
http://www.bom.gov.au/oceanography/projects/abslmp/data/data.shtml
}
\description{
Bureau of Meteorology Australian Baseline Sea Level Monitoring Project, with
measurements smoothed to eliminate variability.
}
\details{
\itemize{
\item Day in 2023: time variable in hourly increments
\item Sea Level: metres above Tide Gauge Zero
\item Water Temperature: degrees Celcius
\item Air Temperature: degrees Celcius
\item Barometric Pressure: hPa
}
}
\keyword{datasets}
