% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_SoilNitrogen_Harden}
\alias{ML_SoilNitrogen_Harden}
\title{Soil Mineral Nitrogen with Stubble Management}
\format{
csv file with 28 rows and 3 columns
}
\source{
https://data.csiro.au/collection/csiro:61293?q=population\%20data&_st=keyword&_str=210&_si=7
}
\description{
CSIRO Harden Long-Term Tillage Experiment
}
\details{
\itemize{
\item Year: time variable in sporadic increments
\item Burn: kilograms per hectare of mineral soil nitrogen in the soil profile from 0 to 160 millimetres with stubble burning
\item Bash: kilograms per hectare of mineral soil nitrogen in the soil profile from 0 to 160 millimetres with stubble bashing
}
}
\keyword{datasets}
