% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_SouthernBluefinTuna}
\alias{ML_SouthernBluefinTuna}
\title{Southern Bluefin Tuna}
\format{
csv file with 23 rows and 4 columns
}
\source{
https://doi.org/10.25919/nzrp-6702
}
\description{
CSIRO Fisheries Biomass Provisioning Services
}
\details{
\itemize{
\item Fin Year: time variable for financial year 1 July to 30 June
\item Catch: kilograms caught per year
\item GVP: Gross value product in dollars as market price times catch
\item EV: Exchange value in dollars as the value of ecosystem services as if a market existed
}
}
\keyword{datasets}
