% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_TempsMax_CapeOtway}
\alias{ML_TempsMax_CapeOtway}
\title{Maximum daily temperatures at Cape Otway}
\format{
csv file with 8888 rows and 2 columns
}
\source{
http://www.bom.gov.au/climate/data/?ref=ftr, Station number 090015
}
\description{
Bureau of Meteorology, raw (unhomogenized) data.
}
\details{
\itemize{
\item Year: time variable in daily increments
\item Max: maximum temperature in degrees centigrade
}
}
\keyword{datasets}
