% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_TreeShelterBelts_Tasmania}
\alias{ML_TreeShelterBelts_Tasmania}
\title{Tree shelter belts}
\format{
csv file with 52 rows and 3 columns
}
\description{
Sheep gross margins per paddock with and without trees
}
\details{
\itemize{
\item Year: time variable in annual increments
\item With trees: gross margins per paddock with tree shelter belt
\item Without trees: gross margins per paddock without tree shelter belt
}
}
\author{
Tim Capon \email{tim.capon@csiro.au}
}
\keyword{datasets}
