% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_WaterFarmDams_Riverina}
\alias{ML_WaterFarmDams_Riverina}
\title{Water in Farm Dams in the Riverina}
\format{
csv file with 19313 rows and 3 columns
}
\description{
Simulated water volumes with and without tree windbreaks to control evaporation
and a merino sheep flock size of 1500.
}
\details{
\itemize{
\item Day: time variable in daily increments
\item Without: water volumes in m3 without a windbreak
\item With: water volumes in m3 with a windbreak
}
}
\author{
Tim Capon \email{tim.capon@csiro.au}
}
\keyword{datasets}
