% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_WheatCanolaSheep_Narrendera}
\alias{ML_WheatCanolaSheep_Narrendera}
\title{Farm Adaptation at Narrendera, New South Wales}
\format{
csv file with 50 rows and 5 columns
}
\source{
https://nccarf.edu.au/will-primary-producers-continue-adjust-practices-and-technologies-change-production/
}
\description{
APSIM generated gross margins for wheat (W), canola (C) and sheep (S)
}
\details{
\itemize{
\item Year: time variable in annual increments
\item CWWCWW GM: Gross margins per hectare for CWWCWW rotation
\item CWWSSS GM: Gross margins per hectare for CWWSSS rotation
\item SSSSC GM: Gross margins per hectare for SSSSC rotation
\item S GM: Gross margins per hectare for continuous S
}
}
\keyword{datasets}
