% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{ML_WheatFutures_KansasCity}
\alias{ML_WheatFutures_KansasCity}
\title{Kansas City Hard Red Wheat Futures}
\format{
csv file with 253 rows and 7 columns
}
\source{
https://finance.yahoo.com/quote (search for KE=F)
}
\description{
Daily futures from 25 July 2023 to 25 January 2024
}
\details{
\itemize{
\item Day: time variable in days since 25 July 2023
\item Open: Price at market open in USD/ton
\item High: Daily high price in USD/ton
\item Low: Daily low price in USD/ton
\item Close: Price at market close in USD/ton
\item Adj Close: Same as Close for futures
\item Volume: Daily number of trades
}
}
\keyword{datasets}
