% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessHertz.R
\name{myReadData}
\alias{myReadData}
\title{Function to read and return data}
\usage{
myReadData(file)
}
\value{
dataframe
}
\description{
Read internal data and return dataframe
}
\examples{
  myReadData("MyData")
}
