% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessHertz.R
\name{OUPDataRead}
\alias{OUPDataRead}
\title{Function to read and return data}
\usage{
OUPDataRead(file = "MyData")
}
\value{
dataframe
}
\description{
Read internal data and return dataframe
}
\examples{
  OUPDataRead("MyData")
}
