% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessHertz.R
\name{OUPHelp}
\alias{OUPHelp}
\title{Function to launch roxygen and HTML help for GregsOUP}
\usage{
OUPHelp()
}
\value{
Help and Browser instances
}
\description{
Launch ?OUProcess, ?Analytical, ?MaximumLikelihood and ?MonteCarlo
in RStudio and OUP_Help in the default browser.
}
\examples{
  OUPHelp()
}
