% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessHertz.R
\name{OUPReadData}
\alias{OUPReadData}
\title{Function to read and return data}
\usage{
OUPReadData(file)
}
\value{
dataframe
}
\description{
Read internal data and return dataframe
}
\examples{
  OUPReadData("MyData")
}
