% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessHertz.R
\name{OUPRibbonHelp}
\alias{OUPRibbonHelp}
\title{Function to launch Ribbon Help for GregsOUPR6}
\usage{
OUPRibbonHelp()
}
\value{
Browser instance
}
\description{
Launch OUP_Help in the default browser.
}
\examples{
  OUPRibbonHelp()
}
