% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUProcessHertz.R
\name{OUProcess}
\alias{OUProcess}
\title{R6 class implementing a container to synchronize OUP objects.}
\description{
The Analytical, FiniteDifference, MaximumLikelihood and MonteCarlo objects
are a complete set of functions for maximum likelihood estimation and
for the calculation of probabilities, option prices, visiting times, first
passage times, decision thresholds and more--everything for a Real Options
Analysis.  Each object can be used by itself.  This object, OUProcess, will
instantiate and synchronize the other objects.
}
\section{Usage:}{
This object, OUProcess, and the Analytical, FiniteDifference,
MaximumLikelihood and MonteCarlo objects can be instantiated together
by:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  OUP <- OUProcess$new()
}\if{html}{\out{</div>}}

Then pointers to individual objects can be accessed by:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  A <- OUP$get_Analytical()
  FD <- OUP$get_FiniteDifference()
  ML <- OUP$get_MaximumLikelihood()
  MC <- OUP$get_MonteCarlo()
}\if{html}{\out{</div>}}

OUP has other public methods, but these are called by individual objects
and won't do anything if called by a user.
}

\section{Discussion}{
OUP will have pointers to A, FD, ML and MC.  A, FD, ML and MC will each
have a pointer to OUP. Thus A, FD, ML and MC can follow the pointers to
synchronize with each other. For example, rho, mu and sigma estimated in ML
will be propagated to the other objects:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  ML$Estimates()
  A$Option()
}\if{html}{\out{</div>}}

Alternately, each object can be instantiated individually:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  A <- Analytical$new()
  FD <- FiniteDifference$new()
  ML <- MaximumLikelihood$new()
  MC <- MonteCarlo$new()
}\if{html}{\out{</div>}}

In this case, A, FD, ML and MC will not have pointers and will not
synchronize. To estimate rho, mu and sigma and calculate option prices
would require:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  estimates <- ML$Estimates()
  rho <- estimates$rho
  mu <- estimates$mu
  sigma <- estimates$sigma
  A$Option(rho=rho,mu=mu,sigma=sigma)
}\if{html}{\out{</div>}}

Upon initialization, OUP will ask if it is being instantiated by an RShiny
app.  If so, it will store the session pointer and use it to coordinate
clipboard access between RShiny and the A, FD, ML and MC objects.
}

\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-OUProcess-initialize}{\code{OUProcess$new()}}
    \item \href{#method-OUProcess-get_Analytical}{\code{OUProcess$get_Analytical()}}
    \item \href{#method-OUProcess-get_FiniteDifference}{\code{OUProcess$get_FiniteDifference()}}
    \item \href{#method-OUProcess-get_MaximumLikelihood}{\code{OUProcess$get_MaximumLikelihood()}}
    \item \href{#method-OUProcess-get_MonteCarlo}{\code{OUProcess$get_MonteCarlo()}}
    \item \href{#method-OUProcess-send_oup_params}{\code{OUProcess$send_oup_params()}}
    \item \href{#method-OUProcess-send_y_stoch_args}{\code{OUProcess$send_y_stoch_args()}}
    \item \href{#method-OUProcess-send_x_stoch_args}{\code{OUProcess$send_x_stoch_args()}}
    \item \href{#method-OUProcess-send_t_stoch_args}{\code{OUProcess$send_t_stoch_args()}}
    \item \href{#method-OUProcess-send_plot_args}{\code{OUProcess$send_plot_args()}}
    \item \href{#method-OUProcess-send_plot_info}{\code{OUProcess$send_plot_info()}}
    \item \href{#method-OUProcess-send_flags}{\code{OUProcess$send_flags()}}
    \item \href{#method-OUProcess-CopyToClipboard}{\code{OUProcess$CopyToClipboard()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-initialize}{}}}
\subsection{\code{OUProcess$new()}}{
  Create an OUProcess object as a container for other objects
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$new(session = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{session}}{session pointer to RShiny app}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new OUProcess object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-get_Analytical"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-get_Analytical}{}}}
\subsection{\code{OUProcess$get_Analytical()}}{
  Pointer to Analytical object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$get_Analytical()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    pointer
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-get_FiniteDifference"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-get_FiniteDifference}{}}}
\subsection{\code{OUProcess$get_FiniteDifference()}}{
  Pointer to FiniteDifference object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$get_FiniteDifference()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    pointer
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-get_MaximumLikelihood"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-get_MaximumLikelihood}{}}}
\subsection{\code{OUProcess$get_MaximumLikelihood()}}{
  Pointer to MaximumLikelihood object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$get_MaximumLikelihood()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    pointer
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-get_MonteCarlo"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-get_MonteCarlo}{}}}
\subsection{\code{OUProcess$get_MonteCarlo()}}{
  Pointer to MonteCarlo object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$get_MonteCarlo()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    pointer
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-send_oup_params"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-send_oup_params}{}}}
\subsection{\code{OUProcess$send_oup_params()}}{
  Function called by modules to coordinate parameters
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$send_oup_params(rho = NULL, mu = NULL, sigma = NULL, who = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rho}}{rate parameter 0<=rho<inf}
      \item{\code{mu}}{location parameter -inf<mu<inf}
      \item{\code{sigma}}{scale parameter -inf<sigma<inf}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-send_y_stoch_args"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-send_y_stoch_args}{}}}
\subsection{\code{OUProcess$send_y_stoch_args()}}{
  Function called by modules to coordinate arguments for y as a stochastic state
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$send_y_stoch_args(
  t = NULL,
  y = NULL,
  x = NULL,
  psi = NULL,
  who = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{t}}{vector of m times -inf<t<inf}
      \item{\code{y}}{vector of n states -inf<y<inf}
      \item{\code{x}}{initial state -inf<x<inf}
      \item{\code{psi}}{<=0 for integral -inf to y, >0 for integral y to inf}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-send_x_stoch_args"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-send_x_stoch_args}{}}}
\subsection{\code{OUProcess$send_x_stoch_args()}}{
  Function called by modules to coordinate x as a stochastic state and its arguments
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$send_x_stoch_args(
  s = NULL,
  x = NULL,
  y = NULL,
  r = NULL,
  phi = NULL,
  who = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{s}}{vector of m times -inf<s<t}
      \item{\code{x}}{vector of n states -inf<x<inf}
      \item{\code{y}}{terminal state -inf<y<inf}
      \item{\code{r}}{discount rate -inf<r<inf}
      \item{\code{phi}}{<=0 for exit option, >0 for entry option}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-send_t_stoch_args"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-send_t_stoch_args}{}}}
\subsection{\code{OUProcess$send_t_stoch_args()}}{
  Function called by modules to coordinate t stochastic arguments
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$send_t_stoch_args(
  t = NULL,
  k = NULL,
  x = NULL,
  omega = NULL,
  Ppct = NULL,
  who = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{t}}{vector of m times -inf<t<inf}
      \item{\code{k}}{threshold -inf<k<inf}
      \item{\code{x}}{initial state -inf<x<inf}
      \item{\code{omega}}{degree of irreversibility 0<=omega<=1}
      \item{\code{Ppct}}{passage time probability for a percentile  0.01<=Ppct<=0.99}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-send_plot_args"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-send_plot_args}{}}}
\subsection{\code{OUProcess$send_plot_args()}}{
  Function called by modules to coordinate arguments for plotting
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$send_plot_args(pmax = NULL, ptmax = NULL, who = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pmax}}{maximum transition density}
      \item{\code{ptmax}}{maximum visiting time and first passage time densities}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-send_plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-send_plot_info}{}}}
\subsection{\code{OUProcess$send_plot_info()}}{
  Function called by modules to coordinate information for plotting
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$send_plot_info(
  fontfamily = NULL,
  fontsize = NULL,
  fileformat = NULL,
  filewidth = NULL,
  fileheight = NULL,
  theme = NULL,
  opaque = NULL,
  walls = NULL,
  floor = NULL,
  labels = NULL,
  who = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fontfamily}}{font family for plot labels}
      \item{\code{fontsize}}{font size for plot labels}
      \item{\code{fileformat}}{'png' or 'svg'}
      \item{\code{filewidth}}{pixel width of 2D plot, pixel width and height of 3D plot}
      \item{\code{fileheight}}{pixel height of 2D plot}
      \item{\code{theme}}{'light' or 'dark'}
      \item{\code{opaque}}{transparent to opaque background 0.0<=opaque<=1.0}
      \item{\code{walls}}{3D walls TRUE or FALSE}
      \item{\code{floor}}{3D floor TRUE or FALSE}
      \item{\code{labels}}{title and parameters TRUE or FALSE}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-send_flags"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-send_flags}{}}}
\subsection{\code{OUProcess$send_flags()}}{
  Function called by modules to coordinate flags for plotting and copying
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$send_flags(plotit = NULL, copyit = NULL, who = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plotit}}{automatic plot after calculation TRUE or FALSE}
      \item{\code{copyit}}{copy to clipboard TRUE or FALSE}
      \item{\code{who}}{object id of sender}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OUProcess-CopyToClipboard"></a>}}
\if{latex}{\out{\hypertarget{method-OUProcess-CopyToClipboard}{}}}
\subsection{\code{OUProcess$CopyToClipboard()}}{
  Write to clipboard
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OUProcess$CopyToClipboard(clip)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{clip}}{data frame, matrix or vector}
    }
    \if{html}{\out{</div>}}
  }
}

}
