% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Agric_NSW_FarmDamsRiverina}
\alias{Agric_NSW_FarmDamsRiverina}
\title{Water in Farm Dams in the Riverina of New South Wales}
\format{
csv file with 19313 rows and 4 columns
}
\source{
corresponding author: Tim Capon \email{tim.capon.csiro.au}
}
\description{
Stock water management to improve drought resilience in intensive-use grazing landscapes.
MatLab simulated water volumes for a 4,000 cubic metre dam.
}
\details{
\itemize{
\item Run Days: time variable in daily increments
\item Baseline: water volumes in cubic metres without a windbreak
\item TwentyPct scenario: water volumes in cubic metres with a windbreak to suppress evaporation by 20\%
\item FortyPct scenario: water volumes in cubic metres with a windbreak to suppress evaporation by 40\%
}
}
\author{
Helena Clayton, Sally Thompson, Tim Capon, Greg Hertzler, Philip Graham, and David Lindenmayer, 2026
}
\keyword{datasets}
