% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Agric_NSW_FarmDamsRiverina}
\alias{Agric_NSW_FarmDamsRiverina}
\title{Water in Farm Dams in the Riverina of New South Wales}
\format{
csv file with 19313 rows and 3 columns
}
\description{
Matlab generated water volumes in a 4000 cubic metre dam, with and without
tree windbreaks to control evaporation, and a merino sheep flock size of
1500 dry sheep equivalents.
}
\details{
\itemize{
\item Day: time variable in daily increments
\item Without: water volumes in cubic metres without a windbreak
\item With: water volumes in cubic metres with a windbreak
}
}
\author{
{ Tim Capon \email{tim.capon.csiro.au}, Helena Clayton \email{helena.clayton@anu.edu.au}, Sally Thompson \email{sally.thompson@uwa.edu.au}, Greg Hertzler \email{ghertzlerau@gmail.com}, Philip Graham \email{phil@gramadvisory.com.au}, David Lindemayer \email{david.lindemayer@anu.edu} }
}
\keyword{datasets}
