% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Agric_NSW_SoilHealthHarden}
\alias{Agric_NSW_SoilHealthHarden}
\title{Soil Health with Stubble Management}
\format{
csv file with 30 rows and 5 columns
}
\source{
https://data.csiro.au/collection/csiro:61293?q=population\%20data&_st=keyword&_str=210&_si=7
}
\description{
CSIRO Harden Long-Term Tillage Experiment
}
\details{
\itemize{
\item Year: time variable in sporadic increments
\item N Burn: kilograms per hectare of mineral soil nitrogen in the soil profile from 0 to 160 millimetres with stubble burning
\item N Bash: kilograms per hectare of mineral soil nitrogen in the soil profile from 0 to 160 millimetres with stubble bashing
\item W Burn: millimetres of water in the soil profile from 0 to 160 millimetres with stubble burning
\item W Bash: millimetres of water in the soil profile from 0 to 160 millimetres with stubble bashing
}
}
\keyword{datasets}
