% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Agric_SA_GMHawker}
\alias{Agric_SA_GMHawker}
\title{Farm Adaptation at Hawker, South Australia}
\format{
csv file with 108 rows and 7 columns
}
\source{
https://nccarf.edu.au/will-primary-producers-continue-adjust-practices-and-technologies-change-production/
}
\description{
APSIM generated Wheat yields and manually simulated dry-sheep equivalents
}
\details{
\itemize{
\item Year: time variable in annual increments
\item Annual rain: Annual rainfall in millimetres
\item Apr-Oct rain: Growing season rainfall in millimetres
\item Wheat Yld: Wheat yield in tonnes per hectare
\item Sheep DSE: Dry-sheep equivalents per hectare
\item Wheat GM: Wheat gross margins per hectare
\item Sheep GM: Sheep gross margins per hectare
}
}
\keyword{datasets}
