% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Climate_SeaLevel_PortKembla}
\alias{Climate_SeaLevel_PortKembla}
\title{Sea Level at Port Kembla}
\format{
csv file with 7985 rows and 5 columns
}
\source{
http://www.bom.gov.au/oceanography/projects/abslmp/data/data.shtml
}
\description{
Australian Baseline Sea Level Monitoring Project.
}
\details{
\itemize{
\item Day in 2023: time variable in hourly increments
\item Sea Level: metres above Tide Gauge Zero
\item Water Temperature: degrees Celsius
\item Air Temperature: degrees Celsius
\item Barometric Pressure: hPa
}

Sea levels above Tide Guage Zero for 2023.

This is radically smoothed data.  Do not use this data for decisions under uncertainty.
}
\author{
Bureau of Meteorology
}
\keyword{datasets}
