% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Climate_Sunspots}
\alias{Climate_Sunspots}
\title{Sunspot Numbers}
\format{
csv file with 3316 rows and 2 columns
}
\source{
https://doi.org/10.24414/qnza-ac80
}
\description{
International Sunspot Number V2.0.
}
\details{
\itemize{
\item DecimalYear: time variable in daily increments
\item Average: Simple average of the daily total sunspot number over all days of each calendar month.
}

Sunspot numbers and groups of sunspot numbers, averaged by month.
}
\author{
Royal Observatory of Belgium
}
\keyword{datasets}
