% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Climate_TempsMax_TennantCreek}
\alias{Climate_TempsMax_TennantCreek}
\title{Maximum daily temperatures at Tennant Creek}
\format{
csv file with 9115 rows and 2 columns
}
\source{
http://www.bom.gov.au/climate/data/?ref=ftr, Station number 015135
}
\description{
Raw (Unhomogenized) data.
}
\details{
\itemize{
\item Year: time variable in daily increments
\item Max: maximum temperature in degrees centigrade
}
}
\author{
Bureau of Meteorology
}
\keyword{datasets}
