% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Default}
\alias{Default}
\title{Default}
\format{
csv file with 176 rows and 2 columns
}
\description{
Simulated data matching default data in Maximum Likelihood object
}
\details{
\itemize{
\item tau: time variable
\item z: state variable
}
}
\keyword{datasets}
