% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Ecosys_Albatross}
\alias{Ecosys_Albatross}
\title{Albatross Egg Counts}
\format{
csv file with 60 rows and 4 columns
}
\source{
https://tsx.org.au/tsx2024
}
\description{
The Australian Threatened Species Index, 2024.
Nesting egg counts of three Albatross species in Tasmania.
}
\details{
\itemize{
\item Year: time variable in annual increments
\item Wandering Albatross: egg count
\item Black-browed Albatross: egg count
\item Grey-headed Albatross: egg count
}
}
\author{
Australian Government, Department of Climate Change, Energy, the Environment and Water
}
\keyword{datasets}
