% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Ecosys_IrrigationWaterSupply}
\alias{Ecosys_IrrigationWaterSupply}
\title{Water Supply for Irrigated Agriculture in Eastern Australia}
\format{
csv file with 13 rows and 2 columns
}
\source{
https://doi.org/10.25919/7jj7-8826
}
\description{
CSIRO Agricultural water supply data for national ecosystem services accounts.
}
\details{
\itemize{
\item Fin Year: time variable for financial year 1 July to 30 June
\item WaterSupply: irrigation water supplied in Australia in megalitres
}
}
\keyword{datasets}
