% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Ecosys_Kangaroos}
\alias{Ecosys_Kangaroos}
\title{Kangaroo Population and Harvest in South Australia}
\format{
csv file with 46 rows and 7 columns
}
\source{
https://www.environment.sa.gov.au/topics/animals-and-plants/sustainable-use-of-animals-and-plants/kangaroo-conservation-and-management/quotas-harvest-data
}
\description{
Government of South Australia, Department of Environment, population surveys and
meat processor records of harvest aggregated across all regions of South Australia.
}
\details{
\itemize{
\item Year: time variable
\item Red Pop: Estimated population of Red Kangaroos
\item Red Harvest: Number of Red Kangaroos commercially harvested
\item Grey Pop: Estimated population of Western Grey Kangaroos
\item Grey Harvest: Number of Western Grey Kangaroos commercially harvested
\item Euro Pop: Estimated population of European Kangaroos
\item Euro Harvest: Number of European Kangaroos commercially harvested
}
}
\keyword{datasets}
