% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Ecosys_SydneyWater}
\alias{Ecosys_SydneyWater}
\title{Sydney Drinking Water Catchment}
\format{
csv file with 120 rows and 14 columns
}
\source{
https://waterinsights.waternsw.com.au/12964-sydney-drinking-water-catchment/storage
}
\description{
WaterNSW, WaterInsight.
}
\details{
\itemize{
\item Day: time variable for days from August 2015
\item All: All Storage volume in gigalitres
\item Blue Mtns: Blue Mountains Dams volume in gigalitres
\item Nepean: Nepean Dam volume in gigalitres
\item Avon: Avon Dam volume in gigalitres
\item Wingecarribe: Wingecarribe Reservoir volume in gigalitres
\item Cordeaux: Cordeaux Dam volume in gigalitres
\item Cataract: Cataract Dam volume in gigalitres
\item Warragamba: Warragamba Dam volume in gigalitres
\item Woronora: Woronora Dam volume in gigalitres
\item Prospect: Prospect Reservoir volume in gigalitres
\item Tallowa: Tallowa Dam volume in gigalitres
\item Fitzroy: Fitzroy Falls Dam volume in gigalitres
\item Year: time variable as decimal year
}

Drinking Water Storage from August 2015 to July 2025.
}
\author{
WaterNSW
}
\keyword{datasets}
