% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Ecosys_TropicalRockLobsters}
\alias{Ecosys_TropicalRockLobsters}
\title{Tropical Rock Lobster in Australian Waters}
\format{
csv file with 22 rows and 4 columns
}
\source{
https://doi.org/10.25919/nzrp-6702
}
\description{
National data for Australia's ecosystem services: fisheries biomass provisioning services (2000-01 to 2020-21). v9. CSIRO. Data Collection.
Tropical Rock Lobster aggregated for all of Australia
}
\details{
\itemize{
\item Fin Year: time variable for financial year 1 July to 30 June
\item Catch: kilograms caught per year
\item GVP: Gross value product in dollars as market price times catch
\item EV: Exchange value in dollars as the value of ecosystem services as if a market existed
}
}
\author{
Pascoe, Sean; Liu, Ning; Scheufele, Gabriela; Tetreault Campbell, Sally; & Schmidt, Becky, 2024
}
\keyword{datasets}
