% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{Finance_USDAUD}
\alias{Finance_USDAUD}
\title{US Dollars per Australian Dollar}
\format{
csv file with 258 rows and 7 columns
}
\source{
https://finance.yahoo.com/quote (search for AUDUSD=X)
}
\description{
Daily exchange rates from 1 July 2024 to 30 June 2025,
with AUD as the base currency and USD as the quote currency
}
\details{
\itemize{
\item Day: time variable in days since 1 July 2024
\item Open: Rate at session open in USD/AUD
\item High: Daily high rate in USD/AUD
\item Low: Daily low rate in USD/AUD
\item Close: Rate at session close in USD/AUD
\item Adj Close: Same as Close for exchange rates
\item Year: time variable as decimal year
}
}
\keyword{datasets}
