% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloHertz.R
\name{MonteCarlo}
\alias{MonteCarlo}
\title{R6 Class implementing Monte Carlo simulation}
\description{
This object is a placeholder to be completed in a future version.
}
\section{Methods:}{
To be completed.
}

\section{Plots:}{
To be completed,
}

\section{Arguments of functions:}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  All arguments are optional in all functions.
}\if{html}{\out{</div>}}

To be completed.
}

\section{Using the methods:}{
Demonstration scripts are in files in the 'demo' directory. Identify a
formula and in the console type:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  demo(MC_FormulaName), or
  demo(MC_PlotFormulaName).
}\if{html}{\out{</div>}}
}

\section{Discussion:}{
To be completed.
}

\examples{

## ------------------------------------------------
## Method `MonteCarlo$new`
## ------------------------------------------------

  MC <- MonteCarlo$new()
  MC$get_oup_params()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MonteCarlo-new}{\code{MonteCarlo$new()}}
\item \href{#method-MonteCarlo-set_oup_params}{\code{MonteCarlo$set_oup_params()}}
\item \href{#method-MonteCarlo-set_plot_info}{\code{MonteCarlo$set_plot_info()}}
\item \href{#method-MonteCarlo-get_oup_params}{\code{MonteCarlo$get_oup_params()}}
\item \href{#method-MonteCarlo-get_plot_info}{\code{MonteCarlo$get_plot_info()}}
\item \href{#method-MonteCarlo-get_plot_colors}{\code{MonteCarlo$get_plot_colors()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonteCarlo-new"></a>}}
\if{latex}{\out{\hypertarget{method-MonteCarlo-new}{}}}
\subsection{Method \code{new()}}{
Create a MonteCarlo object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonteCarlo$new(OUP = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OUP}}{pointer set by the OUProcess object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new MonteCarlo object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  MC <- MonteCarlo$new()
  MC$get_oup_params()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonteCarlo-set_oup_params"></a>}}
\if{latex}{\out{\hypertarget{method-MonteCarlo-set_oup_params}{}}}
\subsection{Method \code{set_oup_params()}}{
Set OUP parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonteCarlo$set_oup_params(rho = NULL, mu = NULL, sigma = NULL, who = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rho}}{rate parameter 0<=rho<inf}

\item{\code{mu}}{location parameter -inf<mu<inf}

\item{\code{sigma}}{scale parameter -inf<sigma<inf}

\item{\code{who}}{identifier for object sending the parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list(rho,mu,sigma)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonteCarlo-set_plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-MonteCarlo-set_plot_info}{}}}
\subsection{Method \code{set_plot_info()}}{
Set information for plotting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonteCarlo$set_plot_info(
  type = NULL,
  pmax = NULL,
  ptmax = NULL,
  fontfamily = NULL,
  fontsize = NULL,
  fileformat = NULL,
  filewidth = NULL,
  fileheight = NULL,
  theme = NULL,
  opaque = NULL,
  walls = NULL,
  floor = NULL,
  labels = NULL,
  who = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{= 1 and 2 for 2D, 3 and 4 for 3D}

\item{\code{pmax}}{maximum vertical axis for transition density plot}

\item{\code{ptmax}}{maximum vertical axis for passage time density plot}

\item{\code{fontfamily}}{font family for plot labels}

\item{\code{fontsize}}{font size for plot labels}

\item{\code{fileformat}}{'png' or 'svg'}

\item{\code{filewidth}}{pixel width of 2D plot, pixel width and height of 3D plot}

\item{\code{fileheight}}{pixel height of 2D plot}

\item{\code{theme}}{'light' or 'dark'}

\item{\code{opaque}}{transparent to opaque background 0.0<=opaque<=1.0}

\item{\code{walls}}{3D walls TRUE or FALSE}

\item{\code{floor}}{3D floor TRUE or FALSE}

\item{\code{labels}}{title and parameters TRUE or FALSE}

\item{\code{who}}{identifier for object sending the parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list(type,font,file,theme,3D,labels)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonteCarlo-get_oup_params"></a>}}
\if{latex}{\out{\hypertarget{method-MonteCarlo-get_oup_params}{}}}
\subsection{Method \code{get_oup_params()}}{
Get OUP parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonteCarlo$get_oup_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list(rho,mu,sigma)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonteCarlo-get_plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-MonteCarlo-get_plot_info}{}}}
\subsection{Method \code{get_plot_info()}}{
Get information for plotting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonteCarlo$get_plot_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list(type,font,file,theme,3D,labels)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonteCarlo-get_plot_colors"></a>}}
\if{latex}{\out{\hypertarget{method-MonteCarlo-get_plot_colors}{}}}
\subsection{Method \code{get_plot_colors()}}{
Get colors for plotting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonteCarlo$get_plot_colors()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list(red,ylw,grn,cyn,blu,mgn,gry,background,font,reverse)
}
}
}
