% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptionalPackages}
\alias{OptionalPackages}
\alias{RcppParallelInstalled}
\alias{RcppsitmoInstalled}
\title{Optional packages for parallel processing}
\usage{
RcppParallelInstalled()

RcppsitmoInstalled()
}
\value{
bool <- RcppParallelInstalled()

bool <- RcppsitmoInstalled()
}
\description{
Queries whether functions are compiled with RcppParallel or fall back
to Rcpp only.  Also queries whether random numbers are generated
by RcppParallel using sitmo() or fall back to Rcpp using rnorm().  And
there are functions for setting and getting the number of threads.
}
\section{Discussion}{
Rcpp calculates hundreds of times faster than R6 objects.  RcppParallel
calculates five to eight times faster than Rcpp on a typical laptop with
12 threads, and thousands of times faster than R6 objects.  For Monte Carlo
simulations, RcppParallel using sitmo() generates random numbers six times
faster Rcpp using rnorm() and nine times faster R6 using rnorm().  Both
RcppParallel and sitmo are optional but recommended:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ install.packages("RcppParallel", "sitmo")
}\if{html}{\out{</div>}}
}

