% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{OUP_ObservationInterval}
\alias{OUP_ObservationInterval}
\title{Observation intervals and the Ornstein-Uhlenbeck Process}
\format{
csv file with 366 rows and 3 columns
}
\description{
Monte-Carlo simulation to demonstrate the effect of the observation interval on the
rate of convergence, rho, and the scale, sigma.
}
\details{
\itemize{
\item Day: time variable in daily increments
\item z: sample paths with daily observations
\item Year: time variable as decimal year
}

Observation intervals are measured in two different units:  days and years.

Both sets of estimates have the same rho(t-s), where rho is the rate of convergence and
t-s is the observation interval.  For observation intervals measured in days, t-s=1 and
rho=1/2.  For observation intervals measured in years, t-s=1/365 and rho=365/2.  Location
parameter, mu, is the same in both sets of estimates.  Scale parameter, sigma, is larger
if rho is larger because the asymptotic variance, sigma^2/2rho, is the same in both sets
of estimates.  If rho is 365 times bigger, sigma is 365^0.5 = 19.105 times bigger.

Because rho(t-s) is the same, alpha = 0.5(1+exp(-rho(t-s))) is also the same, where alpha
identifies the probability distribution of the estimates.
}
\keyword{datasets}
