% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihoodHertz.R
\docType{data}
\name{OUP_SampleSize}
\alias{OUP_SampleSize}
\title{Sample sizes for the Ornstein-Uhlenbeck Process}
\format{
csv file with 501 rows and 4 columns
}
\description{
Monte-Carlo simulation to demonstrate the effect of sample size on discovering the
parameters of the Ornstein-Uhlenbeck Process.
}
\details{
\itemize{
\item year: time variable in annual increments for all sample paths
\item small: 5 observations
\item medium: 50 observations
\item large: 500 observations
}

The first 5 and the first 50 observations are the same as those of 500 simulated
observations.  More observations get closer to the true parameters of rho=0.5,
mu=15 and sigma=50.
}
\keyword{datasets}
