% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHertz.R
\docType{data}
\name{OUP_StationaryIncrements}
\alias{OUP_StationaryIncrements}
\title{Stationary increments}
\format{
csv file with 1001 rows and 5 columns
}
\description{
Stationary increments are created from pseudo-random standard normal variables generated by
four random number generators.
}
\details{
\itemize{
\item tau: time variable
\item dqrng: dqrng::pcg64 uniform random numbers with Ziggurat transform to normal
\item mt19937: std::mt19937 uniform random numbers with Polar Box-Muller transform to normal
\item sitmo:  sitmo::prng uniform random numbers with Polar Box-Muller transform to normal
\item rnorm:  rnorm uniform random numbers with cumulative normal inversion
}

An important assumption of the Ornstein-Uhlenbeck Process is stationary increments. An
increment is eps(t-s)^0.5, where eps~N(0,1) and t-s is a time interval.  This data has
time intervals equal to one and eps from dqrng, an optional package, mt19937, the generator
in c++, sitmo, an optional package, and rnorm, the default generator in R.

Random number generators must pass many tests.  Here we test for stationary increments as
a special case of the Ornstein-Uhlenbeck Process.  The Ornstein-Uhlenbeck Process is
stationary if rho goes to infinity and sigma goes to infinity in such a way that
sigma/(2rho)^0.5 goes to a constant.

To conduct a test, first estimate the unrestricted parameters.  With time intervals equal
to one, parameter rho equal to 3 or more is closer to infinity than to zero.  Then impose
restrictions. Choose a large number with an easy square root, such as 1600.  Set 2rho=1600,
or rho=800. Set mu=0.  Set sigma=40, which is the square root of 1600 and imposes the
restriction that sigma/(2rho)^0.5=1.

Then do a Likelihood Ratio Test comparing the unrestricted and restricted estimates.  If
the Log Likelihoods are not significantly different, the hypothesis of stationary increments
cannot be rejected.
}
\keyword{datasets}
